requirements.txt
aiodns==3.5.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
attrs==25.3.0
blinker==1.9.0
ccxt==4.4.95
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cryptography==45.0.5
Flask==3.1.1
frozenlist==1.7.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
multidict==6.6.3
numpy==2.3.1
pandas==2.3.1
propcache==0.3.2
pycares==4.9.0
pycparser==2.22
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
requests==2.32.4
setuptools==80.9.0
six==1.17.0
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
Werkzeug==3.1.3
yarl==1.20.1
slack-sdk=3.36.0